backend/requirements.txt
fastapi==0.135.2
passlib==1.7.4
protobuf==7.34.1
pydantic==2.12.5
pydantic_settings==2.13.1
python-dotenv==1.2.2
python_jose==3.5.0
python_jose==3.3.0
SQLAlchemy==2.0.36
uvicorn==0.42.0
